+++ POSCoreTests/CalendarPlanLogic/CalendarWorkCreatorTests/CreateCalendarWorkShould.cs
@@ -24,7 +24,7 @@ namespace POSCoreTests.CalendarPlanLogic.CalendarWorkCreatorTests
         {
             var estimateWork = new EstimateWork("ЭЛЕКТРОХИМИЧЕСКАЯ ЗАЩИТА", 0, 0, 0, 0);
 
-            var calendarWork = _calendarWorkCreator.CreateCalendarWork(DateTime.Today, estimateWork, new int[] { 0 });
+            var calendarWork = _calendarWorkCreator.CreateCalendarWork(DateTime.Today, estimateWork, new decimal[] { 0 });
 
             Assert.NotNull(calendarWork);
         }
@@ -34,7 +34,7 @@ namespace POSCoreTests.CalendarPlanLogic.CalendarWorkCreatorTests
         {
             var estimateWork = new EstimateWork("ЭЛЕКТРОХИМИЧЕСКАЯ ЗАЩИТА", 0, 0, 0, 0);
 
-            var calendarWork = _calendarWorkCreator.CreateCalendarWork(DateTime.Today, estimateWork, new int[] { 0 });
+            var calendarWork = _calendarWorkCreator.CreateCalendarWork(DateTime.Today, estimateWork, new decimal[] { 0 });
 
             Assert.AreEqual(estimateWork.WorkName, calendarWork.WorkName);
         }
@@ -44,7 +44,7 @@ namespace POSCoreTests.CalendarPlanLogic.CalendarWorkCreatorTests
         {
             var estimateWork = new EstimateWork("ЭЛЕКТРОХИМИЧЕСКАЯ ЗАЩИТА", 0, 0, 0, 0);
 
-            var calendarWork = _calendarWorkCreator.CreateCalendarWork(DateTime.Today, estimateWork, new int[] { 0 });
+            var calendarWork = _calendarWorkCreator.CreateCalendarWork(DateTime.Today, estimateWork, new decimal[] { 0 });
 
             Assert.AreEqual(estimateWork.TotalCost, calendarWork.TotalCost);
         }
@@ -54,7 +54,7 @@ namespace POSCoreTests.CalendarPlanLogic.CalendarWorkCreatorTests
         {
             var estimateWork = new EstimateWork("ЭЛЕКТРОХИМИЧЕСКАЯ ЗАЩИТА", (decimal)2.027, (decimal)1.541, (decimal)55.464, 0);
 
-            var calendarWork = _calendarWorkCreator.CreateCalendarWork(DateTime.Today, estimateWork, new int[] { 0 });
+            var calendarWork = _calendarWorkCreator.CreateCalendarWork(DateTime.Today, estimateWork, new decimal[] { 0 });
 
             Assert.AreEqual(estimateWork.TotalCost - estimateWork.EquipmentCost - estimateWork.OtherProductsCost, calendarWork.TotalCostIncludingContructionAndInstallationWorks);
         }
@@ -64,7 +64,7 @@ namespace POSCoreTests.CalendarPlanLogic.CalendarWorkCreatorTests
         {
             var estimateWork = new EstimateWork("ЭЛЕКТРОХИМИЧЕСКАЯ ЗАЩИТА", 0, 0, 0, 1);
 
-            var calendarWork = _calendarWorkCreator.CreateCalendarWork(DateTime.Today, estimateWork, new int[] { 0 });
+            var calendarWork = _calendarWorkCreator.CreateCalendarWork(DateTime.Today, estimateWork, new decimal[] { 0 });
 
             Assert.AreEqual(estimateWork.Chapter, calendarWork.EstimateChapter);
         }
@@ -74,7 +74,7 @@ namespace POSCoreTests.CalendarPlanLogic.CalendarWorkCreatorTests
         {
             var estimateWork = new EstimateWork("ЭЛЕКТРОХИМИЧЕСКАЯ ЗАЩИТА", 0, 0, (decimal)11.324, 0);
             var initialDate = new DateTime(1999, 9, 21);
-            var percentages = new int[] { 1 };
+            var percentages = new decimal[] { 1 };
 
             var calendarWork = _calendarWorkCreator.CreateCalendarWork(initialDate, estimateWork, percentages);
